@@ -1,14 +1,14 @@
-import React from 'react';
-import { Link } from 'react-router-dom'
+import React from "react";
+import { Link } from "react-router-dom";
 
 const Header = () => {
-    return (
-        <header className="col-12 col-md-8">
-            <Link to={'/'}>
-                <h1 className="text-center">React blog</h1>
-            </Link>
-        </header>
-    )
-}
+  return (
+    <header className="col-12 col-md-8 mt-2">
+      <Link to={"/"}>
+        <h1 className="text-center ">React blog</h1>
+      </Link>
+    </header>
+  );
+};
 
 export default Header;
